shared memory results added into readme.md

+++ shared_memory/mmap_sem_shm/shm_writer.c
@@ -15,12 +15,13 @@
 #include "mmap_shm.h"
 
 struct timeval start, end;
-const unsigned long int memSize = ( SHM_ROW_SIZE * SHM_COL_SIZE );
+const unsigned long long int memSize = ( SHM_ROW_SIZE * SHM_COL_SIZE );
 
 // 
 /*
  *  Lorem Ipsum
  */
+// 
 int 
 main() {
     int fd, fdSync, fdTime;
